Update usageMetrics logic
