+++ subgraphs/stakedao/src/modules/Metrics.ts
@@ -7,7 +7,7 @@ import {
 } from "../../generated/schema";
 import * as utils from "../common/utils";
 import * as constants from "../common/constants";
-import { BigInt, Address } from "@graphprotocol/graph-ts";
+import { BigInt, Address, ethereum } from "@graphprotocol/graph-ts";
 
 export function updateFinancials(blockNumber: BigInt, timestamp: BigInt): void {
   // Number of days since Unix epoch
@@ -41,36 +41,27 @@ export function updateFinancials(blockNumber: BigInt, timestamp: BigInt): void {
   financialMetrics.save();
 }
 
-export function updateUsageMetrics(
-  blockNumber: BigInt,
-  timestamp: BigInt,
-  from: Address
-): void {
+export function updateUsageMetrics(block: ethereum.Block, from: Address): void {
   // Number of days since Unix epoch
-  let id: i64 = timestamp.toI64() / constants.SECONDS_PER_DAY;
-  let usageMetrics = UsageMetricsDailySnapshot.load(id.toString());
-
-  if (!usageMetrics) {
-    usageMetrics = new UsageMetricsDailySnapshot(id.toString());
-    usageMetrics.protocol = constants.ETHEREUM_PROTOCOL_ID;
-
-    usageMetrics.activeUsers = 0;
-    usageMetrics.totalUniqueUsers = 0;
-    usageMetrics.dailyTransactionCount = 0;
-  }
+  let id: i64 = block.timestamp.toI64() / constants.SECONDS_PER_DAY;
+  let usageMetrics = utils.getOrCreateUsageMetricSnapshot(block);
 
   // Update the block number and timestamp to that of the last transaction of that day
-  usageMetrics.blockNumber = blockNumber;
-  usageMetrics.timestamp = timestamp;
+  usageMetrics.blockNumber = block.number;
+  usageMetrics.timestamp = block.timestamp;
   usageMetrics.dailyTransactionCount += 1;
 
   let accountId = from.toHexString();
   let account = _Account.load(accountId);
+  let protocol = utils.getOrCreateYieldAggregator(constants.ETHEREUM_PROTOCOL_ID);
   if (!account) {
     account = new _Account(accountId);
     account.save();
-    usageMetrics.totalUniqueUsers += 1;
+
+    protocol.totalUniqueUsers += 1;
+    protocol.save();
   }
+  usageMetrics.totalUniqueUsers = protocol.totalUniqueUsers;
 
   // Combine the id and the user address to generate a unique user id for the day
   let dailyActiveAccountId = id.toString() + "-" + from.toHexString();
